Move logic to process source within validate method
